Questão Corrigida

+++ Q07.py
@@ -1,7 +1,7 @@
 nome = input("Nome do estudante: ")
 materia = input("Qual é a matéria?: ")
-n1 = int(input("Digite sua primeira nota: "))
-n2 = int(input("Digite sua segunda nota: "))
+n1 = float(input("Digite sua primeira nota: "))
+n2 = float(input("Digite sua segunda nota: "))
 m = (n1 + n2)/2
 if m >= 6:
     situacao = "aprovado"
